onpoint_pay/requirements.txt
Django==5.0.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
django-cors-headers==4.3.1
django-filter==23.3
celery==5.3.4
redis==5.0.1
python-decouple==3.8
requests==2.31.0
cryptography==41.0.7
gunicorn==21.2.0
whitenoise==6.6.0

# Development dependencies
django-debug-toolbar==4.2.0
django-extensions==3.2.3
